earnchain-backend/requirements.txt
Flask==2.3.2
python-telegram-bot==20.4
python-dotenv==1.0.0